finished article editor
this commit breaks a lot of thing.

we have now the ability to upload images. and they will automatically be arranged in folders in the backend. This required a change on other pages (like index, category), and Database default value as we don't store the relative path of the file anymore but only the filename.